Clear VRAM maps and sprites on reset
Both maps their attributes, and sprites are now cleared on reset, and
the window is also hidden. This ensures that no leftover tiles remain
after a reset, and the screen turns blank immediately.

As part of this:
- Fix VRAMClearRoutine, so that it actually works (but bc needs to be
set to bc + $0101; handled by VRAMClear macro)
- Add an EnsureVBlank subroutine, which ensures that we’re in a
V-Blank. The previous EnsureVBlank subroutine (which checks for V-Blank
or H-Blank) has been renamed EnsureVRAMAccess.
- The game will also now start in Double Speed mode.