Don't dynamically enable/disable of double speed
When trying to compensate for the timer change, subtle bugs could
result on an actual GBA due to timing differences. This would
eventually cause the timer to be set incorrectly.

To compensate, the timer adjustment code has been stripped away, and
the game now remains in double speed mode whilst it is being run.

Could remove the DisableDoubleSpeed subroutine if there is no need for
it.